init: add gromacs example
This runs caliper metric extraction on the gromacs
repository. It is interesting to see a few inflection
points with huge changes! I think next I want to figure out
a way to run gromacs without GPU, and then build different
versions into a Dockerfile bast to test. The alternative
is using Dockerfile they have provided, but I just pulled
the 2018 variant and it was built requiring GPU (which
I do not have handy)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>